move some methods elsewhere
