+++ ephemeris/shed_tools.py
@@ -42,14 +42,13 @@ from collections import namedtuple
 import yaml
 from bioblend.galaxy.client import ConnectionError
 from bioblend.galaxy.toolshed import ToolShedClient
-from bioblend.toolshed import ToolShedInstance
 from galaxy.tools.verify.interactor import GalaxyInteractorApi, verify_tool
 
 from . import get_galaxy_connection, load_yaml_file
 from .ephemeris_log import disable_external_library_logging, setup_global_logger
 from .get_tool_list_from_galaxy import GiToToolYaml, the_same_repository, tools_for_repository
 from .shed_tools_args import parser
-
+from .shed_tools_methods import *
 
 class InstallToolManager(object):
     """Manages the installation of new tools on a galaxy instance"""
@@ -77,7 +76,7 @@ class InstallToolManager(object):
         if check_revision:
             # If we want to check if revisions are equal, flatten the list,
             # so each tool - revision combination has its own entry
-            installed_repos = _flatten_repo_info(self.installed_tools())
+            installed_repos = flatten_repo_info(self.installed_tools())
         else:
             # If we do not care about revision equality, do not do the flatten
             # action to limit the number of comparisons.
@@ -111,7 +110,7 @@ class InstallToolManager(object):
         counter = 0
 
         # Start by flattening the repo list per revision
-        flattened_repos = _flatten_repo_info(tools)
+        flattened_repos = flatten_repo_info(tools)
         total_num_repositories = len(flattened_repos)
 
         # Complete the repo information, and make sure each tool has a revision
@@ -252,7 +251,7 @@ class InstallToolManager(object):
             # something like installed_repository_revisions(self.gi)
             tools = self.installed_tools()
 
-        target_repositories = _flatten_repo_info(tools)
+        target_repositories = flatten_repo_info(tools)
 
         installed_tools = []
         for target_repository in target_repositories:
@@ -374,59 +373,6 @@ class InstallToolManager(object):
         return False
 
 
-def complete_repo_information(tool, default_toolshed_url, require_tool_panel_info, default_install_tool_dependencies,
-                              default_install_repository_dependencies,
-                              default_install_resolver_dependencies, force_latest_revision):
-    repo = dict()
-    # We need those values. Throw a KeyError when not present
-    repo['name'] = tool['name']
-    repo['owner'] = tool['owner']
-    repo['tool_panel_section_id'] = tool.get('tool_panel_section_id')
-    repo['tool_panel_section_label'] = tool.get('tool_panel_section_label')
-    if require_tool_panel_info and repo['tool_panel_section_id'] is None and repo[
-            'tool_panel_section_label'] is None and 'data_manager' not in repo.get('name'):
-        raise KeyError("Either tool_panel_section_id or tool_panel_section_name must be defined for tool '{0}'.".format(
-            repo.get('name')))
-    repo['tool_shed_url'] = format_tool_shed_url(tool.get('tool_shed_url', default_toolshed_url))
-    repo['changeset_revision'] = tool.get('changeset_revision')
-    repo = get_changeset_revisions(repo, force_latest_revision)
-    repo['install_repository_dependencies'] = tool.get('install_repository_dependencies',
-                                                       default_install_repository_dependencies)
-    repo['install_resolver_dependencies'] = tool.get('install_resolver_dependencies',
-                                                     default_install_resolver_dependencies)
-    repo['install_tool_dependencies'] = tool.get('install_tool_dependencies', default_install_tool_dependencies)
-    return repo
-
-
-def format_tool_shed_url(tool_shed_url):
-    formatted_tool_shed_url = tool_shed_url
-    if not formatted_tool_shed_url.endswith('/'):
-        formatted_tool_shed_url += '/'
-    if not formatted_tool_shed_url.startswith('http'):
-        formatted_tool_shed_url = 'https://' + formatted_tool_shed_url
-    return formatted_tool_shed_url
-
-
-def get_changeset_revisions(repository, force_latest_revision=False):
-    """
-    Select the correct changeset revision for a repository,
-    and make sure the repository exists
-    (i.e a request to the tool shed with name and owner returns a list of revisions).
-    Return repository or None, if the repository could not be found on the specified tool shed.
-    """
-    # Do not connect to the internet when not necessary
-    if repository.get('changeset_revision') is None or force_latest_revision:
-        ts = ToolShedInstance(url=repository['tool_shed_url'])
-        # Get the set revision or set it to the latest installable revision
-        installable_revisions = ts.repositories.get_ordered_installable_revisions(repository['name'],
-                                                                                  repository['owner'])
-        if not installable_revisions:  #
-            raise LookupError("Repo does not exist in tool shed: {0}".format(repository))
-        repository['changeset_revision'] = installable_revisions[-1]
-
-    return repository
-
-
 def log_repository_install_error(repository, start, msg, log):
     """
     Log failed tool installations. Return a dictionary wiyh information
@@ -479,39 +425,6 @@ def log_repository_install_start(repository, counter, total_num_repositories, in
             dt.datetime.now() - installation_start
         )
     )
-
-
-def _flatten_repo_info(repositories):
-    """
-    Flatten the dict containing info about what tools to install.
-    The tool definition YAML file allows multiple revisions to be listed for
-    the same tool. To enable simple, iterative processing of the info in this
-    script, flatten the `tools_info` list to include one entry per tool revision.
-    :type repositories: list of dicts
-    :param repositories: Each dict in this list should contain info about a tool.
-    :rtype: list of dicts
-    :return: Return a list of dicts that correspond to the input argument such
-             that if an input element contained `revisions` key with multiple
-             values, those will be returned as separate list items.
-    """
-    flattened_list = []
-    for repo_info in repositories:
-        if 'revisions' in repo_info:
-            revisions = repo_info.get('revisions', [])
-            repo_info.pop('revisions', None)  # Set default to avoid key error  when removing revisions
-            if not revisions:  # Revisions are empty list or None
-                flattened_list.append(repo_info)
-            else:
-                for revision in revisions:
-                    # A new dictionary must be created, otherwise there will
-                    # be aliasing of dictionaries. Which leads to multiple
-                    # repos with the same revision in the end result.
-                    new_repo_info = dict(**repo_info)
-                    new_repo_info['changeset_revision'] = revision
-                    flattened_list.append(new_repo_info)
-        else:  # Revision was not defined at all
-            flattened_list.append(repo_info)
-    return flattened_list
 
 
 def main():
